# optimize SQL query in collectAZAllocationStats()